+++ zyte_common_items/items.py
@@ -111,6 +111,12 @@ class ArticleFromList(Item):
 
 @attrs.define(kw_only=True)
 class Article(Item):
+    """Article, typically seen on online news websites, blogs, or announcement
+    sections.
+
+    :attr:`url` is the only required attribute.
+    """
+
     #: Headline or title.
     headline: Optional[str] = None
 
@@ -261,7 +267,8 @@ class ArticleList(Item):
 class ProductVariant(Item):
     """:class:`Product` variant.
 
-    See :attr:`Product.variants`.
+    See :attr:`Product.variants`, :class:`ProductVariantExtractor`,
+    :class:`ProductVariantSelectorExtractor`.
     """
 
     #: List of name-value pais of data about a specific, otherwise unmapped
@@ -406,8 +413,7 @@ class ProductVariant(Item):
 class Product(Item):
     """Product from an e-commerce website.
 
-    The :attr:`url` attribute is the only required attribute, all other fields
-    are optional.
+    :attr:`url` is the only required attribute.
     """
 
     #: List of name-value pais of data about a specific, otherwise unmapped
@@ -633,7 +639,8 @@ class ProductFromList(Item):
     """Product from a product list from a product listing page of an e-commerce
     webpage.
 
-    See :class:`ProductList`.
+    See :class:`ProductList`, :class:`ProductFromListExtractor`,
+    :class:`ProductFromListSelectorExtractor`.
     """
 
     #: Price currency `ISO 4217`_ alphabetic code (e.g. ``"USD"``).
@@ -700,8 +707,7 @@ class ProductList(Item):
 
     It represents, for example, a single page from a category.
 
-    The :attr:`url` attribute is the only required attribute, all other fields
-    are optional.
+    :attr:`url` is the only required attribute.
     """
 
     #: Webpage `breadcrumb trail`_.
@@ -756,7 +762,10 @@ class ProductList(Item):
 @attrs.define(slots=True, kw_only=True)
 class BusinessPlace(Item):
     """Business place, with properties typically seen on maps or business
-    listings."""
+    listings.
+
+    :attr:`url` is the only required attribute.
+    """
 
     #: Unique identifier of the place on the website.
     placeId: Optional[str] = None
@@ -855,6 +864,12 @@ class BusinessPlace(Item):
 
 @attrs.define(slots=True, kw_only=True)
 class RealEstate(Item):
+    """Real state offer, typically seen on real estate offer aggregator
+    websites.
+
+    :attr:`url` is the only required attribute.
+    """
+
     #: The url of the final response, after any redirects.
     url: str = attrs.field(converter=url_to_str)
 
@@ -955,6 +970,9 @@ class RealEstate(Item):
 
 
 class RequestListCaster:
+    """attrs converter to turn lists of :class:`Request` instances into lists
+    of :class:`ProbabilityRequest` instances."""
+
     def __init__(self, target):
         self._target = target
 
@@ -1050,6 +1068,12 @@ class ArticleNavigation(Item):
 
 @attrs.define(kw_only=True)
 class JobPosting(Item):
+    """A job posting, typically seen on job posting websites or websites of
+    companies that are hiring.
+
+    :attr:`url` is the only required attribute.
+    """
+
     #: The url of the final response, after any redirects.
     url: str = attrs.field(converter=url_to_str)