+++ zyte_common_items/processors.py
@@ -8,8 +8,17 @@ from parsel import Selector, SelectorList
 from price_parser import Price
 from web_poet.mixins import ResponseShortcutsMixin
 from zyte_parsers import Breadcrumb as zp_Breadcrumb
-from zyte_parsers import extract_brand_name, extract_breadcrumbs, extract_price
+from zyte_parsers import Gtin as zp_Gtin
+from zyte_parsers import (
+    extract_brand_name,
+    extract_breadcrumbs,
+    extract_gtin,
+    extract_price,
+    extract_rating,
+    extract_review_count,
+)
 
+from . import AggregateRating, Gtin
 from .items import Breadcrumb
 
 
@@ -43,6 +52,9 @@ def _format_price(price: Price) -> Optional[str]:
 def only_handle_nodes(
     f: Callable[[Union[Selector, HtmlElement], Any], Any]
 ) -> Callable[[Any, Any], Any]:
+    """Decorator for processors that only runs a decorated processor if the
+    input is of type :class:`Selector` or :class:`HtmlElement`."""
+
     @wraps(f)
     def wrapper(value: Any, page: Any) -> Any:
         value = _handle_selectorlist(value)
@@ -184,3 +196,93 @@ def description_processor(value: Any, page: Any) -> Any:
     page._description_node = cleaned_node
     page._description_str = cleaned_text
     return cleaned_text
+
+
+def gtin_processor(
+    value: Union[SelectorList, Selector, HtmlElement, str], page: Any
+) -> Any:
+    """Convert the data into a list of :class:`~zyte_common_items.Gtin` objects if possible.
+
+    Supported inputs are :class:`str`, :class:`~parsel.selector.Selector`,
+    :class:`~parsel.selector.SelectorList`, :class:`~lxml.html.HtmlElement`, an
+    iterable of :class:`str` and an iterable of :class:`zyte_parsers.Gtin`
+    objects.
+    Other inputs are returned as is.
+    """
+
+    def _from_zp_gtin(zp_value: zp_Gtin) -> Gtin:
+        return Gtin(type=zp_value.type, value=zp_value.value)
+
+    results = []
+    if isinstance(value, SelectorList):
+        for sel in value:
+            if result := extract_gtin(sel):
+                results.append(_from_zp_gtin(result))
+    elif isinstance(value, (Selector, HtmlElement, str)):
+        if result := extract_gtin(value):
+            results.append(_from_zp_gtin(result))
+    elif isinstance(value, Iterable):
+        for item in value:
+            if isinstance(item, zp_Gtin):
+                results.append(_from_zp_gtin(item))
+            elif isinstance(item, str):
+                results.append(_from_zp_gtin(extract_gtin(item)))
+            else:
+                results.append(item)
+    else:
+        return value
+    return results or None
+
+
+def rating_processor(value: Any, page: Any) -> Any:
+    """Convert the data into an :class:`~zyte_common_items.AggregateRating` object if possible.
+
+    Supported inputs are :class:`~parsel.selector.Selector`,
+    :class:`~parsel.selector.SelectorList` and :class:`~lxml.html.HtmlElement`.
+    The input can also be a dictionary with one or more of the following keys:
+    "ratingValue", "bestRating", "reviewCount". The values for those keys will
+    be assigned to the respective result fields. If the value for "ratingValue"
+    is a :class:`~parsel.selector.Selector`,
+    :class:`~parsel.selector.SelectorList` or :class:`~lxml.html.HtmlElement`
+    instance, the final values for the ratingValue and bestRating fields will
+    be extracted from it (if the "bestRating" key is also present in the
+    dictionary it will take precedence). If the value for the "reviewCount" key
+    is of one of those types, the final value for the reviewCount field will be
+    extracted from it.
+    Other inputs are returned as is.
+    """
+    value = _handle_selectorlist(value)
+    if isinstance(value, (Selector, HtmlElement)):
+        zp_rating = extract_rating(value)
+        result = AggregateRating(
+            reviewCount=extract_review_count(value),
+            bestRating=zp_rating.bestRating,
+            ratingValue=zp_rating.ratingValue,
+        )
+        if result.reviewCount or result.bestRating or result.ratingValue:
+            return result
+        return None
+    elif isinstance(value, dict):
+        result = AggregateRating()
+
+        review_count = _handle_selectorlist(value.get("reviewCount"))
+        if isinstance(review_count, (Selector, HtmlElement)):
+            result.reviewCount = extract_review_count(review_count)
+        elif review_count is not None:
+            result.reviewCount = int(review_count)
+
+        rating_value = _handle_selectorlist(value.get("ratingValue"))
+        if isinstance(rating_value, (Selector, HtmlElement)):
+            zp_rating = extract_rating(rating_value)
+            result.ratingValue = zp_rating.ratingValue
+            result.bestRating = zp_rating.bestRating
+        elif rating_value is not None:
+            result.ratingValue = float(rating_value)
+
+        if (best_rating := value.get("bestRating")) is not None:
+            result.bestRating = float(best_rating)
+
+        if result.reviewCount or result.bestRating or result.ratingValue:
+            return result
+        return None
+    return value