+++ CHANGELOG.rst
@@ -2,12 +2,32 @@
 Changelog
 =========
 
+0.14.0 (2024-01-16)
+===================
+
+* Now requires ``zyte-parsers >= 0.4.0``.
+
+* Added :func:`zyte_common_items.processors.gtin_processor` and enabled it for
+  the ``gtin`` fields in the page classes for
+  :class:`~zyte_common_items.Product`.
+
+* Improved the API documentation.
+
+0.13.0 (2023-11-09)
+===================
+
+* Added ``Auto``-prefixed versions of :ref:`page objects <page-objects>`, such
+  as :meth:`~zyte_common_items.AutoProductPage`, that return data from Zyte API
+  automatic extraction from their fields by default, and can be used to more
+  easily override that data with custom parsing logic.
+
+
 0.12.0 (2023-10-27)
 ===================
 
 * Added :meth:`~zyte_common_items.base.ProbabilityMixin.get_probability` helper
   method in item classes (e.g. :class:`~zyte_common_items.Product`,
-  :class:`~zyte_common_items.Article`) and 
+  :class:`~zyte_common_items.Article`) and
   :class:`~zyte_common_items.ProbabilityRequest`.